nilfs2: correct logic of translation virtual blocks number into physical ones
The issue report is contained description of mount failure:

[  822.390075] NILFS version 2 loaded
[  822.436888] segctord starting. Construction interval = 5 seconds, CP frequency < 30 seconds
[  822.437107] NILFS: corrupt root inode.

Investigation of the issue discovered that it takes places
reading of ifile's block from not proper placement because of
successful ending of operation of virtual block number translation
into physical one.

This patch adds checking of start checkpoint and end checkpoint by
comparing with current checkpoint number during translation of
virtual block number into physical block number. Because it makes
sense to operate with blocks that are alive for current checkpoint.

As a result, mount fails with such error message:

[17348.793335] NILFS warning (device loop0): nilfs_ifile_get_inode_block: unable to read inode: 2
[17348.793869] NILFS: get root inode failed

Reported-by: Bryce Gibson <bryce@gibson-consulting.com.au>
Signed-off-by: Vyacheslav Dubeyko <slava@dubeyko.com>
CC: Ryusuke Konishi <konishi.ryusuke@lab.ntt.co.jp>